+++ GraphicsEngine/GraphicsEngine/Content/IndexBuffer.cpp
@@ -3,7 +3,7 @@
 
 using namespace GraphicsEngine;
 
-IndexBuffer::IndexBuffer(const D3DBase& d3dBase, const void* indices, uint32_t indexCount, uint32_t indexByteSize, DXGI_FORMAT indexFormat) :
+IndexBuffer::IndexBuffer(const D3DBase& d3dBase, const void* indices, size_t indexCount, size_t indexByteSize, DXGI_FORMAT indexFormat) :
 	m_indexFormat(indexFormat),
 	BaseGeometryBuffer(d3dBase, indices, indexCount, indexByteSize)
 {
